pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiohttp>=3.11.16",
    "cryptography>=44.0.2",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "starlette>=0.46.2",
    "uvicorn>=0.34.1",
]
